finish https://app.pluralsight.com/library/courses/laravel-php-framework-getting-started-users-authentication/discussion

+++ app/Http/Controllers/PostController.php
@@ -5,8 +5,12 @@ namespace App\Http\Controllers;
 use App\Like;
 use App\Post;
 use App\Tag;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Session\Store;
+use Illuminate\Support\Facades\Auth;
+use Reflector;
+
 class PostController extends Controller
 {
     //
@@ -34,11 +38,17 @@ class PostController extends Controller
 
     public function getAdminEdit ($id) {
         $post = Post::find($id);
+        if (Gate::denies('update-post', $post)) {
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
     }
     public function getAdminDelete ($id) {
         $post = Post::find($id);
+        if (Gate::denies('update-post', $post)) {
+            return redirect()->back();
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
@@ -57,11 +67,15 @@ class PostController extends Controller
             'title' => 'required|min:6',
             'content' => 'required|min:10',
         ]);
+        $user = Auth::user();
+        // dd($user->id);
 
         $post = new Post([
             'title' => $request->input('title'),
-            'content' => $request->input('content')
+            'content' => $request->input('content'),
+            'user_id' => $user->id
         ]);
+
         $post->save();
         $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
         return redirect()->route('admin.index')->with('info', 'Post create, Title is: ' . $request->input('title'));
@@ -74,6 +88,9 @@ class PostController extends Controller
         ]);
 
         $post = Post::find($request->input(('id')));
+        if (Gate::denies('update-post', $post)) {
+            return redirect()->back();
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
